Fix process management commands on Windows
stop, status, and restart commands were broken on Windows because
processExists() and stopAllWeclaw() used Unix-only APIs (syscall.Signal(0),
syscall.SIGTERM, pkill) without platform build tags.

- Move processExists() and stopAllWeclaw() to platform-specific files
- proc_unix.go: same Unix implementation (//go:build !windows)
- proc_windows.go: use tasklist/taskkill instead of signals/pkill
- restart.go: delegate to stopAllWeclaw() instead of duplicating logic

Fixes #74
