+++ cmd/proc_unix.go
@@ -3,10 +3,37 @@
 package cmd
 
 import (
+	"fmt"
+	"os"
 	"os/exec"
 	"syscall"
+	"time"
 )
 
 func setSysProcAttr(cmd *exec.Cmd) {
 	cmd.SysProcAttr = &syscall.SysProcAttr{Setsid: true}
+}
+
+func processExists(pid int) bool {
+	p, err := os.FindProcess(pid)
+	if err != nil {
+		return false
+	}
+	return p.Signal(syscall.Signal(0)) == nil
+}
+
+func stopAllWeclaw() {
+	if pid, err := readPid(); err == nil && processExists(pid) {
+		if p, err := os.FindProcess(pid); err == nil {
+			_ = p.Signal(syscall.SIGTERM)
+		}
+	}
+	os.Remove(pidFile())
+
+	exe, err := os.Executable()
+	if err != nil {
+		return
+	}
+	_ = exec.Command("pkill", "-f", exe+" start").Run()
+	time.Sleep(500 * time.Millisecond)
 }
