+++ cmd/proc_windows.go
@@ -2,8 +2,30 @@
 
 package cmd
 
-import "os/exec"
+import (
+	"fmt"
+	"os"
+	"os/exec"
+	"strings"
+	"time"
+)
 
 func setSysProcAttr(_ *exec.Cmd) {
-	// No Setsid on Windows — process is already detached via Start()
+}
+
+func processExists(pid int) bool {
+	out, err := exec.Command("tasklist", "/FI", fmt.Sprintf("PID eq %d", pid)).Output()
+	if err != nil {
+		return false
+	}
+	return strings.Contains(string(out), fmt.Sprintf(" %d ", pid))
+}
+
+func stopAllWeclaw() {
+	if pid, err := readPid(); err == nil && processExists(pid) {
+		_ = exec.Command("taskkill", "/F", "/PID", fmt.Sprintf("%d", pid)).Run()
+	}
+	os.Remove(pidFile())
+	_ = exec.Command("taskkill", "/F", "/IM", "weclaw.exe").Run()
+	time.Sleep(500 * time.Millisecond)
 }
