+++ cmd/restart.go
@@ -2,9 +2,6 @@ package cmd
 
 import (
 	"fmt"
-	"os"
-	"syscall"
-	"time"
 
 	"github.com/spf13/cobra"
 )
@@ -17,23 +14,7 @@ var restartCmd = &cobra.Command{
 	Use:   "restart",
 	Short: "Restart the background weclaw process",
 	RunE: func(cmd *cobra.Command, args []string) error {
-		// Stop if running
-		pid, err := readPid()
-		if err == nil && processExists(pid) {
-			fmt.Printf("Stopping weclaw (pid=%d)...\n", pid)
-			if p, err := os.FindProcess(pid); err == nil {
-				p.Signal(syscall.SIGTERM)
-			}
-			for i := 0; i < 20; i++ {
-				if !processExists(pid) {
-					break
-				}
-				time.Sleep(500 * time.Millisecond)
-			}
-			os.Remove(pidFile())
-		}
-
-		// Start
+		stopAllWeclaw()
 		fmt.Println("Starting weclaw...")
 		return runDaemon()
 	},
